# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Artist Directory API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SocialLinks {
  website?: string;
  instagram?: string;
  twitter?: string;
  facebook?: string;
  youtube?: string;
  spotify?: string;
  appleMusic?: string;
  soundcloud?: string;
  tiktok?: string;
  bandcamp?: string;
}

export type AlbumType = (typeof AlbumType)[keyof typeof AlbumType];

export const AlbumType = {
  album: "album",
  ep: "ep",
  single: "single",
  mixtape: "mixtape",
} as const;

export interface Album {
  title: string;
  year: number;
  type: AlbumType;
  spotifyUrl?: string;
  appleMusicUrl?: string;
  youtubeMusicUrl?: string;
  tidalUrl?: string;
  deezerUrl?: string;
  amazonMusicUrl?: string;
}

export interface MusicVideo {
  title: string;
  /** YouTube or Vimeo URL */
  url: string;
  year?: number;
  description?: string;
}

export interface PressQuote {
  quote: string;
  source: string;
  year?: number;
}

export interface Artist {
  id: number;
  slug: string;
  name: string;
  bio: string;
  shortBio?: string;
  genres: string[];
  origin?: string;
  formedYear?: number;
  imageUrl?: string;
  socialLinks?: SocialLinks;
  discography?: Album[];
  musicVideos?: MusicVideo[];
  pressQuotes?: PressQuote[];
  bookingEmail?: string;
  pressEmail?: string;
  labels?: string[];
  members?: string[];
  tags?: string[];
  /** Additional context specifically written for LLM consumption */
  llmContext?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateArtistInput {
  name: string;
  slug?: string;
  bio: string;
  shortBio?: string;
  genres: string[];
  origin?: string;
  formedYear?: number;
  imageUrl?: string;
  socialLinks?: SocialLinks;
  discography?: Album[];
  musicVideos?: MusicVideo[];
  pressQuotes?: PressQuote[];
  bookingEmail?: string;
  pressEmail?: string;
  labels?: string[];
  members?: string[];
  tags?: string[];
  llmContext?: string;
}

export interface UpdateArtistInput {
  name?: string;
  slug?: string;
  bio?: string;
  shortBio?: string;
  genres?: string[];
  origin?: string;
  formedYear?: number;
  imageUrl?: string;
  socialLinks?: SocialLinks;
  discography?: Album[];
  musicVideos?: MusicVideo[];
  pressQuotes?: PressQuote[];
  bookingEmail?: string;
  pressEmail?: string;
  labels?: string[];
  members?: string[];
  tags?: string[];
  llmContext?: string;
}

export interface ArtistList {
  artists: Artist[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface ArtistSearchResults {
  results: Artist[];
  query: string;
  total: number;
}

export interface GenreList {
  genres: string[];
}

export type DirectoryStatsTopGenresItem = {
  genre: string;
  count: number;
};

export interface DirectoryStats {
  totalArtists: number;
  totalGenres: number;
  recentlyAdded: Artist[];
  topGenres: DirectoryStatsTopGenresItem[];
}

export type ListArtistsParams = {
  page?: number;
  limit?: number;
  genre?: string;
  search?: string;
};

export type SearchArtistsParams = {
  q: string;
  genre?: string;
  limit?: number;
};