# lib/api-zod/src/generated/types/album.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Artist Directory API
 * OpenAPI spec version: 0.1.0
 */
import type { AlbumType } from "./albumType";

export interface Album {
  title: string;
  year: number;
  type: AlbumType;
  spotifyUrl?: string;
  appleMusicUrl?: string;
  youtubeMusicUrl?: string;
  tidalUrl?: string;
  deezerUrl?: string;
  amazonMusicUrl?: string;
}