# lib/api-zod/src/generated/types/artistList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Artist Directory API
 * OpenAPI spec version: 0.1.0
 */
import type { Artist } from "./artist";

export interface ArtistList {
  artists: Artist[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}