# lib/api-zod/src/generated/types/createArtistInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Artist Directory API
 * OpenAPI spec version: 0.1.0
 */
import type { Album } from "./album";
import type { MusicVideo } from "./musicVideo";
import type { PressQuote } from "./pressQuote";
import type { SocialLinks } from "./socialLinks";

export interface CreateArtistInput {
  name: string;
  slug?: string;
  bio: string;
  shortBio?: string;
  genres: string[];
  origin?: string;
  formedYear?: number;
  imageUrl?: string;
  socialLinks?: SocialLinks;
  discography?: Album[];
  musicVideos?: MusicVideo[];
  pressQuotes?: PressQuote[];
  bookingEmail?: string;
  pressEmail?: string;
  labels?: string[];
  members?: string[];
  tags?: string[];
  llmContext?: string;
}