# lib/api-zod/src/generated/types/directoryStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Artist Directory API
 * OpenAPI spec version: 0.1.0
 */
import type { Artist } from "./artist";
import type { DirectoryStatsTopGenresItem } from "./directoryStatsTopGenresItem";

export interface DirectoryStats {
  totalArtists: number;
  totalGenres: number;
  recentlyAdded: Artist[];
  topGenres: DirectoryStatsTopGenresItem[];
}