# lib/api-zod/src/generated/types/musicVideo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Artist Directory API
 * OpenAPI spec version: 0.1.0
 */

export interface MusicVideo {
  title: string;
  /** YouTube or Vimeo URL */
  url: string;
  year?: number;
  description?: string;
}