# lib/api-zod/src/generated/types/pressQuote.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Music Artist Directory API
 * OpenAPI spec version: 0.1.0
 */

export interface PressQuote {
  quote: string;
  source: string;
  year?: number;
}